# specs/001-ophir-observability-hub/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ophir Observability Hub API
  version: 0.1.0
  description: Admin, source registry, integration health, dashboard links, and OTLP-compatible ingestion proxy API.
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: Health
  - name: Auth
  - name: Sources
  - name: Integrations
  - name: Dashboards
  - name: Telemetry Ingestion
paths:
  /health/live:
    get:
      tags: [Health]
      summary: Process liveness probe
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /health/ready:
    get:
      tags: [Health]
      summary: Readiness probe for database and required integrations
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
        "503":
          description: Service is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Create an admin session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me:
    get:
      tags: [Auth]
      summary: Return current admin identity
      security:
        - adminBearer: []
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/sources:
    get:
      tags: [Sources]
      summary: List source applications
      security:
        - adminBearer: []
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SourceApplication"
    post:
      tags: [Sources]
      summary: Register a source application
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceApplication"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/sources/{sourceId}:
    get:
      tags: [Sources]
      summary: Get a source application
      security:
        - adminBearer: []
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceApplication"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Sources]
      summary: Update a source application
      security:
        - adminBearer: []
      parameters:
        - $ref: "#/components/parameters/SourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceRequest"
      responses:
        "200":
          description: Updated source application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceApplication"
  /v1/sources/{sourceId}/credentials:
    post:
      tags: [Sources]
      summary: Create a source ingestion credential
      description: Plaintext key is returned only once.
      security:
        - adminBearer: []
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "201":
          description: Source credential created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSourceCredentialResponse"
  /v1/integrations:
    get:
      tags: [Integrations]
      summary: List telemetry integrations and current health
      security:
        - adminBearer: []
      responses:
        "200":
          description: Integration list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TelemetryIntegration"
  /v1/dashboard-links:
    get:
      tags: [Dashboards]
      summary: List protected Grafana dashboard links
      security:
        - adminBearer: []
      parameters:
        - name: sourceId
          in: query
          required: false
          schema:
            type: string
        - name: telemetryType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TelemetryType"
      responses:
        "200":
          description: Dashboard links
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DashboardLink"
  /otel/v1/logs:
    post:
      tags: [Telemetry Ingestion]
      summary: Accept OTLP-compatible logs and forward to Collector
      security:
        - sourceKey: []
      requestBody:
        required: true
        content:
          application/x-protobuf:
            schema:
              type: string
              format: binary
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          $ref: "#/components/responses/TelemetryAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/CollectorUnavailable"
  /otel/v1/traces:
    post:
      tags: [Telemetry Ingestion]
      summary: Accept OTLP-compatible traces and forward to Collector
      security:
        - sourceKey: []
      requestBody:
        required: true
        content:
          application/x-protobuf:
            schema:
              type: string
              format: binary
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          $ref: "#/components/responses/TelemetryAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/CollectorUnavailable"
  /otel/v1/metrics:
    post:
      tags: [Telemetry Ingestion]
      summary: Accept OTLP-compatible metrics and forward to Collector
      security:
        - sourceKey: []
      requestBody:
        required: true
        content:
          application/x-protobuf:
            schema:
              type: string
              format: binary
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          $ref: "#/components/responses/TelemetryAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/CollectorUnavailable"
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sourceKey:
      type: apiKey
      in: header
      name: x-ophir-source-key
  parameters:
    SourceId:
      name: sourceId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    CollectorUnavailable:
      description: Telemetry accepted by source auth but could not be forwarded to Collector
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TelemetryAccepted:
      description: Telemetry request accepted for forwarding
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TelemetryAcceptedResponse"
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ReadinessResponse:
      type: object
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [ready, not_ready]
        checks:
          type: array
          items:
            type: object
            required: [name, status]
            properties:
              name:
                type: string
              status:
                type: string
                enum: [healthy, degraded, unavailable]
              message:
                type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
    LoginResponse:
      type: object
      required: [accessToken, admin]
      properties:
        accessToken:
          type: string
        admin:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, email, displayName, role, status]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [admin]
        status:
          type: string
          enum: [active, disabled]
    SourceStatus:
      type: string
      enum: [active, disabled]
    TelemetryType:
      type: string
      enum: [logs, traces, metrics, overview]
    CreateSourceRequest:
      type: object
      required: [slug, displayName, environment]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{1,62}$"
        displayName:
          type: string
          minLength: 1
        environment:
          type: string
        ownerName:
          type: string
        ownerContact:
          type: string
    UpdateSourceRequest:
      type: object
      properties:
        displayName:
          type: string
        ownerName:
          type: string
        ownerContact:
          type: string
        status:
          $ref: "#/components/schemas/SourceStatus"
    SourceApplication:
      type: object
      required: [id, slug, displayName, environment, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        slug:
          type: string
        displayName:
          type: string
        environment:
          type: string
        ownerName:
          type: string
        ownerContact:
          type: string
        status:
          $ref: "#/components/schemas/SourceStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSourceCredentialResponse:
      type: object
      required: [credentialId, keyPrefix, plaintextKey]
      properties:
        credentialId:
          type: string
        keyPrefix:
          type: string
        plaintextKey:
          type: string
          description: Returned once. Store securely.
    TelemetryIntegration:
      type: object
      required: [id, kind, name, baseUrl, status]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [collector, loki, tempo, prometheus, grafana]
        name:
          type: string
        baseUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [healthy, degraded, unavailable, unknown]
        lastCheckedAt:
          type: string
          format: date-time
        lastError:
          type: string
    DashboardLink:
      type: object
      required: [id, title, telemetryType, url]
      properties:
        id:
          type: string
        sourceId:
          type: string
        title:
          type: string
        telemetryType:
          $ref: "#/components/schemas/TelemetryType"
        url:
          type: string
          format: uri
    TelemetryAcceptedResponse:
      type: object
      required: [status, requestId]
      properties:
        status:
          type: string
          enum: [accepted]
        requestId:
          type: string
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
        requestId:
          type: string